Sorted up endpoints for better readability Created new endpoints for super admin Updated success messages
