+++ src/server/api/root.ts
@@ -1,5 +1,11 @@
-import { userRouter } from '~/server/api/routers/user';
 import { createCallerFactory, createTRPCRouter } from '~/server/api/trpc';
+import {
+  authRouter,
+  userCreateRouter,
+  userEditRouter,
+  userGetRouter,
+  userRemoveRouter,
+} from './routers/_index';
 
 /**
  * This is the primary router for your server.
@@ -7,7 +13,11 @@ import { createCallerFactory, createTRPCRouter } from '~/server/api/trpc';
  * All routers added in /api/routers should be manually added here.
  */
 export const appRouter = createTRPCRouter({
-  user: userRouter,
+  userCreate: userCreateRouter,
+  userEdit: userEditRouter,
+  userGet: userGetRouter,
+  userRemove: userRemoveRouter,
+  auth: authRouter,
 });
 
 // export type definition of API
