+++ src/server/api/routers/_helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './all.helpers';
+export * from './error.helpers';
 export * from './user.helpers';
