+++ src/server/api/routers/_index.ts
@@ -1,0 +1,2 @@
+export * from './auth/auth';
+export * from './user/_index';
