+++ src/server/api/routers/_types/_index.ts
@@ -1,0 +1,2 @@
+export * from './user.types'
+export * from './all.types'
