+++ src/server/api/routers/_types/user.types.ts
@@ -1,3 +1,4 @@
-export type JwtPayload = {
+// Jwt payload to get the userId
+export type JwtPayloadProp = {
   userId: string;
-}
+};
