+++ src/server/api/routers/user/_index.ts
@@ -1,0 +1,4 @@
+export * from './user.create'
+export * from './user.edit'
+export * from './user.get'
+export * from './user.remove'
